% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_model_metrics.R
\name{calculate_model_metric}
\alias{calculate_model_metric}
\title{Calculate a specified model metric}
\usage{
calculate_model_metric(
  predictions,
  truth_col,
  predictions_col,
  metric,
  event_level_string = "first"
)
}
\arguments{
\item{predictions}{A data frame or tibble containing model predictions}

\item{metric}{The metric to calculate (e.g., "accuracy", "precision", "recall")}

\item{truth}{A vector containing the ground truth values}

\item{event_level}{The event level for precision and recall calculations. Default is "first".}
}
\value{
The calculated value of the specified metric
}
\description{
This function calculates a specified model metric (e.g., accuracy, precision, recall)
for a given set of predictions and ground truth values.
}
\examples{
calculate_model_metric(wine_test_predictions, wine_test_predictions$quality_category, "accuracy")
calculate_model_metric(wine_test_predictions, wine_test_predictions$quality_category, "precision", event_level = "first")

}
